refactor(frontend): move login and signup view into separate folder
