+++ apps/frontend/app/auth/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginView } from "@/components/composites/views/login-view";
+import { LoginView } from "@/components/composites/views/auth/login-view";
 
 export default function LoginPage() {
     return <LoginView />;
