+++ apps/frontend/app/auth/signup/page.tsx
@@ -1,4 +1,4 @@
-import { SignupView } from "@/components/composites/views/signup-view";
+import { SignupView } from "@/components/composites/views/auth/signup-view";
 
 export default function SignupPage() {
     return <SignupView />;
